requirements.txt
python-telegram-bot==20.5
requests
python-dotenv
